refactor(antigravity): cache CLI regexes, simplify fetch_usage logging, extract tests
Thermo-nuclear review fixes for PR #333:

1. Cache is_agy_cli_command regexes with LazyLock instead of recompiling
   on every call. Matches the canonical pattern in redactor.rs/curl_capture.rs
   and the project rs-lazylock rule.

2. Simplify fetch_usage SourceMode logging: the match-on-source-mode with a
   catch-all arm was redundant branching since auto and cli both resolve to
   the same local probe. Replaced with a single debug log.

3. Extract the test module to tests.rs, bringing mod.rs from 1033 to 758
   lines (under the 1k threshold). Matches the cost_pricing.rs #[path]
   pattern already used in the codebase.
